refactor(checker): split cross_file.rs lowering methods to a child module (2000-line cap)
AgentName: StudioOpus
